requirements.txt
tcod
numpy
pillow

pyglet
requests
setuptools
wheel
cffi